@@ -1,25 +1,28 @@
 buildscript {
     repositories {
         mavenCentral()
+        google()
+        jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.0'
+        classpath 'com.android.tools.build:gradle:3.5.3'
     }
 }
 
-task wrapper(type: Wrapper) {
-    gradleVersion = '3.3'
+wrapper {
+    gradleVersion = '5.6'
 }
 
 apply plugin: 'com.android.library'
 
+
 android {
-    compileSdkVersion 22
-    buildToolsVersion '25.0.0'
+    compileSdkVersion 27
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         minSdkVersion 9
-        targetSdkVersion 22
+        targetSdkVersion 27
         versionCode 1
         versionName "1.0"
     }
@@ -35,7 +38,7 @@ android {
 }
 
 dependencies {
-    compile 'com.android.support:support-v4:22.1.1'
+    implementation 'androidx.core:core:1.0.0-beta01'
 }
 
 
@@ -57,3 +60,8 @@ makeJar.dependsOn(clearJar, build)
 
 apply from: './maven_push.gradle'
 
+repositories {
+    google()
+    mavenCentral()
+    jcenter()
+}
